tidied up reducers
