+++ src/datatable/sortable-table.js
@@ -1,15 +1,12 @@
 import { connect } from 'react-redux';
-import { setSortColumn, setFilters } from './actions';
 import Table from './table';
+import { sortData } from './reducers';
 
 const mapStateToProps = state => ({
-  data: state.data,
-  ascending: state.sort.ascending,
-  column: state.sort.column,
-  schema: state.schema
+  data: sortData(state),
+  schema: state.table.schema
 });
 
 export default connect(
-  mapStateToProps,
-  { setSortColumn, setFilters }
+  mapStateToProps
 )(Table);
